pyproject.toml
[project]
name = "e-commerce-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "authlib>=1.6.11",
    "flask>=3.1.3",
    "flask-caching>=2.3.1",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "stripe>=15.0.1",
]
